agentgateway/requirements.txt
# FastAPI and dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# JWT validation
python-jose[cryptography]==3.3.0
cryptography==41.0.7

# HTTP client
httpx==0.25.2

# AWS SDK
boto3==1.34.0

# YAML parsing
PyYAML==6.0.1

# Logging
python-json-logger==2.0.7

# Circuit breaker
circuitbreaker==1.4.0

# Retry logic
tenacity==8.2.3
